Adds rubocop gem
* fixes a bunch of code based on cops
* adds custom rubocop style

+++ app/services/flow_service.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class FlowService
   require 'transcoder'
 
@@ -17,35 +19,35 @@ class FlowService
   # if it's an mp3, do nothing but return the same file structure.
   def transcode_file
     transcoder = Transcoder.new(
-        file: final_file_path,
-        title: @title,
-        contributor: @contributor
+      file: final_file_path,
+      title: @title,
+      contributor: @contributor
     )
-    duration = transcoder.get_duration
+    duration = transcoder.duration
 
     if File.extname(@filename) == '.wav'
       transcoder.transcode(
-          output: final_flac_path,
-          format: Transcoder::FLAC
+        output: final_flac_path,
+        format: Transcoder::FLAC
       )
       transcoder.transcode(
-          output: mp3_file_path,
-          format: Transcoder::MP3_128
+        output: mp3_file_path,
+        format: Transcoder::MP3_128
       )
       transcoder.transcode(
-          output: he_aac_file_path,
-          format: Transcoder::HE_AAC
+        output: he_aac_file_path,
+        format: Transcoder::HE_AAC
       )
       {
-          final_flac_path: final_flac_path,
-          he_aac_file_path: he_aac_file_path,
-          mp3_file_path: mp3_file_path,
-          duration: duration
+        final_flac_path: final_flac_path,
+        he_aac_file_path: he_aac_file_path,
+        mp3_file_path: mp3_file_path,
+        duration: duration
       }
     else
       {
-          mp3_file_path: final_file_path,
-          duration: duration
+        mp3_file_path: final_file_path,
+        duration: duration
       }
     end
   end
@@ -56,7 +58,7 @@ class FlowService
     # Remove any existing file so that we don't keep appending to an old file.
     FileUtils.rm final_file_path, force: true
     # Open final file in append mode
-    File.open(final_file_path, "a") do |f|
+    File.open(final_file_path, 'a') do |f|
       file_chunks.each do |file_chunk_path|
         # Write each chunk to the permanent file
         f.write File.read(file_chunk_path)
@@ -66,41 +68,41 @@ class FlowService
 
   def clean
     FileUtils.rm_rf chunk_file_directory
-    [final_file_path, final_flac_path, mp3_file_path, he_aac_file_path].each do |file|
-      if File.exist?(file)
-        FileUtils.remove file
-      end
+    [final_file_path, final_flac_path, mp3_file_path, he_aac_file_path]
+      .each do |file|
+      FileUtils.remove file if File.exist?(file)
     end
     true
   end
 
   private
 
-    def chunk_file_directory
-      File.join 'tmp', 'flow', @identifier
-    end
+  def chunk_file_directory
+    File.join 'tmp', 'flow', @identifier
+  end
 
-    def file_chunks
-      Dir["#{chunk_file_directory}/*.part*"].sort_by {|f| f.split(".part")[1].to_i }
-    end
+  def file_chunks
+    Dir["#{chunk_file_directory}/*.part*"]
+      .sort_by { |f| f.split('.part')[1].to_i }
+  end
 
-    def final_file_path
-      File.join final_file_directory, File.basename(@filename)
-    end
+  def final_file_path
+    File.join final_file_directory, File.basename(@filename)
+  end
 
-    def final_flac_path
-      "#{final_file_path}.flac"
-    end
+  def final_flac_path
+    "#{final_file_path}.flac"
+  end
 
-    def mp3_file_path
-      "#{final_file_path}.128k.mp3"
-    end
+  def mp3_file_path
+    "#{final_file_path}.128k.mp3"
+  end
 
-    def he_aac_file_path
-      "#{final_file_path}.he-aac.m4a"
-    end
+  def he_aac_file_path
+    "#{final_file_path}.he-aac.m4a"
+  end
 
-    def final_file_directory
-        File.join "tmp", "final"
-    end
+  def final_file_directory
+    File.join 'tmp', 'final'
+  end
 end
